+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## **In-dev** - [0.8.1]
+### Added:
+- `SchemaGenerator::definitions_mut()` which returns a mutable reference to the generator's schema definitions
+
+### Changed:
+- Minimum supported rust version is now 1.37.0
+- Deriving JsonSchema on enums now sets `additionalProperties` to false on generated schemas wherever serde doesn't accept unknown properties. This includes non-unit variants of externally tagged enums, and struct-style variants of all enums that have the `deny_unknown_fields` attribute.
+
+### Fixed
+- Fix use of `#[serde(transparent)]` in combination with `#[schemars(with = ...)]` (https://github.com/GREsau/schemars/pull/67)
+
 ## [0.8.0] - 2020-09-27
 ### Added:
 - `visit::Visitor`, a trait for updating a schema and all schemas it contains recursively. A `SchemaSettings` can now contain a list of visitors.